PR suggestions: avoided repetition in LOGs, changed int to float for CopyFromTexture arguments

+++ Band/Source/Band/Private/BandBlueprintLibrary.cpp
@@ -79,7 +79,7 @@ EBandStatus UBandBlueprintLibrary::CopyFromBuffer(UPARAM(ref) UBandTensor* Tenso
 	return Tensor->CopyFromBuffer(Buffer);
 }
 
-EBandStatus UBandBlueprintLibrary::CopyFromTexture(UPARAM(ref) UBandTensor* Tensor, UTexture2D* Texture, int32 Mean, int32 Std)
+EBandStatus UBandBlueprintLibrary::CopyFromTexture(UPARAM(ref) UBandTensor* Tensor, UTexture2D* Texture, float Mean, float Std)
 {
 	return Tensor->CopyFromTexture(Texture, Mean, Std);
 }
